// src/types/models/economic-operator.ts
export type EconomicOperator = {
  Active: number;
  EO_A_Info: string;
  EO_Address: string;
  EO_Address_City: string;
  EO_Address_Name: string;
  EO_Address_PostCode: string;
  EO_Address_StreetOne: string;
  EO_Address_StreetTwo: string;
  EO_CODE: string;
  EO_CountryReg: string;
  EO_Email: string;
  EO_ExciseNumber1: number;
  EO_ExciseNumber2: string;
  EO_ID: string;
  EO_Name1: string;
  EO_Name2: string;
  EO_OtherID: string;
  EO_Phone: string;
  EO_Type: number;
  Extensibility: string;
  OtherEOID_N_list: string;
  OtherEOID_R: number;
  Reg_3RD: number;
  Reg_EOID: string;
  TAX_N: string;
  VAT_N: string;
  VAT_R: number;
  created_ts: string;
  created_users_id: number;
  last_altered_ts: string;
  last_users_id: string | number;
  used: number;
};